WIP working on C# output
