--- GeneratorSupport/BaseOrThis.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GeneratorSupport
-{
-    public enum BaseOrThis
-    {
-        Neither,
-        Base, 
-        This
-    }
-}
