--- GeneratorSupport/FluentDom/Expressions/Assign.cs
@@ -1,6 +1,0 @@
-﻿namespace GeneratorSupport.FluentDom
-{
-    public class Assign : Expression
-    {
-    }
-}
