--- GeneratorSupport/FluentDom/Expressions/NewObject.cs
@@ -1,14 +1,0 @@
-﻿namespace GeneratorSupport.FluentDom
-{
-    public class NewObject : MethodCall 
-    {
-        private TypeRep typeRep;
-        private string[] arguments;
-
-        public NewObject(TypeRep typeRep, string[] arguments)
-        {
-            this.typeRep = typeRep;
-            this.arguments = arguments;
-        }
-    }
-}
