--- GeneratorSupport/FluentDom/IClassMember.cs
@@ -1,6 +1,0 @@
-﻿namespace GeneratorSupport
-{
-    public interface IClassMember
-    {
-    }
-}
