--- GeneratorSupport/FluentDom/ICtor.cs
@@ -1,6 +1,0 @@
-﻿namespace GeneratorSupport
-{
-    public interface IConstructor
-    {
-    }
-}
