--- GeneratorSupport/FluentDom/IInterface.cs
@@ -1,6 +1,0 @@
-﻿namespace GeneratorSupport.FluentDom
-{
-    public interface IInterface
-    {
-    }
-}
