--- GeneratorSupport/FluentDom/Using.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace GeneratorSupport
-{
-    public class Using
-    {
-        internal static Using Create(string u)
-        {
-            throw new NotImplementedException();
-        }
-
-        internal static Using Create(object u)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
