+++ Samples/SingleLayerCli/CliRootCommandSource.gen.cs
