+++ Samples/SingleLayerCli/SingleLayerCli.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\StarFruit2\StarFruit2.csproj" />
+    <ProjectReference Include="..\..\StarFruit2\StarFruit2.csproj" />
   </ItemGroup>
 
 </Project>
