+++ Samples/TwoLayerCli/TwoLayerCli.csproj
@@ -1,13 +1,14 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
+    <Nullable>enable</Nullable>
     <LangVersion>9.0</LangVersion>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\StarFruit2\StarFruit2.csproj" />
+    <ProjectReference Include="..\..\StarFruit2\StarFruit2.csproj" />
   </ItemGroup>
 
 </Project>
