+++ StarFruit.FluentDomSourceGen/GenerateCommandSource.cs
@@ -1,12 +1,8 @@
-﻿using GeneratorSupport.FluentDom;
+﻿using FluentDom;
 using StarFruit2.Common.Descriptors;
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static GeneratorSupport.FluentDom.Expression;
-//using exp = GeneratorSupport.FluentDom.Expression;
+using static FluentDom.Expression;
 
 // KAD: Extensibility
 // * Use a base/derived overlad extensibility mechanism.Overriding methods may need to navigate DOM
@@ -14,13 +10,14 @@ using static GeneratorSupport.FluentDom.Expression;
 // * Have a separate calculate values method for each method so it can be called at any point
 namespace GeneratorSupport.Tests
 {
-    public class FluentStarFruitTest
+    public class GenerateCommandSource
     {
 
         public void Can_create_code()
         {
             var cli = new CliDescriptor();
             cli.CommandDescriptor = new CommandDescriptor(null, "Fizz", null);
+            var code = CreateCode(cli);
         }
 
         public virtual Code CreateCode(CliDescriptor cli)
@@ -32,7 +29,7 @@ namespace GeneratorSupport.Tests
                     .Usings(
                         "StarFruit2",
                         "System.CommandLine",
-                        "StarFruit2.Common",
+                        new Using("StarFruit2.Common"),
                         "System.CommandLine.Invocation",
                         "System.CommandLine.Parsing")
                     .Class(RootClass(cli.CommandDescriptor))
@@ -163,14 +160,7 @@ namespace GeneratorSupport.Tests
 
     public static class StarFruitExtensions
     {
-        public static TypeRep GetFluentArgumentType(this OptionDescriptor o)
-        => o.Arguments.Any()
-           ? GetFluentArgumentType(o.Arguments.First())
-           : new TypeRep("bool");
-
-        // TODO: Arguments do not currently support generic types
-        public static TypeRep GetFluentArgumentType(this ArgumentDescriptor o)
-          => new TypeRep(o.ArgumentType.TypeAsString());
+   
     }
 }
